♻️ Refactor: 카테고리 포즈 페이지 컴포넌트 분리 및 헤더 컴포넌트  재사용
